Update: change Qubit::Fixed, Token::Integer, Vector.length, MemoryReference.index to usize
This was originally meant to only address #8, changing `Qubit::Fixed`
from `u64` to `usize`. In doing so, the quickest fix was to also update
`Token::Integer` to `usize`, which meant that to make the compiler
happy, I needed to make `Vector.length` and `MemoryReference.index` into
`usize`s too. I'm morally OK with this, as both `length` and `index` to
me _sound_ like they should be of type `usize` instead of `u64`, but I
leave it to the reviewer(s) whether this PR should be merged or burned
with fire.

+++ src/instruction.rs
@@ -162,7 +162,7 @@ impl fmt::Display for ScalarType {
 #[derive(Clone, Debug, Hash, PartialEq)]
 pub struct Vector {
     pub data_type: ScalarType,
-    pub length: u64,
+    pub length: usize,
 }
 
 impl fmt::Display for Vector {
@@ -202,7 +202,7 @@ impl fmt::Display for WaveformInvocation {
 #[derive(Clone, Debug, Hash, PartialEq)]
 pub struct MemoryReference {
     pub name: String,
-    pub index: u64,
+    pub index: usize,
 }
 
 impl Eq for MemoryReference {}
@@ -669,7 +669,7 @@ impl fmt::Display for Instruction {
 
 #[derive(Clone, Debug, Eq, Hash, PartialEq, Serialize, Deserialize)]
 pub enum Qubit {
-    Fixed(u64),
+    Fixed(usize),
     Variable(String),
 }
